abstract extrinsics storage into dedicated class
